requirements.txt
fastapi==0.75.2
pydantic==1.10.22
uvicorn==0.17.6
numpy==1.26.4
requests==2.32.3
pillow==11.2.1
tensorflow==2.19.0
locust
pytest